requirements.txt
plotly-express 
pandas
numpy
seaborn
scikit-learn
imblearn
imbalanced-learn
xgboost
joblib